+++ packages/carp_audio_package/lib/audio.dart
@@ -15,9 +15,7 @@ import 'package:noise_meter/noise_meter.dart';
 import 'package:stats/stats.dart';
 import 'package:permission_handler/permission_handler.dart';
 
-import 'package:carp_core/carp_common/carp_core_common.dart';
-import 'package:carp_core/carp_protocols/carp_core_protocols.dart';
-import 'package:carp_core/carp_data/carp_core_data.dart';
+import 'package:carp_core/carp_core.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 
 part 'audio_probe.dart';