+++ packages/carp_audio_package/lib/audio_probe.dart
@@ -121,7 +121,7 @@ class AudioProbe extends DatumProbe {
       final localApplicationDir = await getApplicationDocumentsDirectory();
       // create a sub-directory for sound files
       final directory = await Directory(
-              '${localApplicationDir.path}/${Settings.CARP_DATA_FILE_PATH}$studyDeploymentPath/$AUDIO_FILES_PATH')
+              '${Settings().deploymentBasePath}/${Settings.CARP_DATA_FILE_PATH}/$AUDIO_FILES_PATH')
           .create(recursive: true);
 
       _path = directory.path;